Move AbsCode and AbsConcept to Map
