+++ lib/Map/AbsCode.js
