+++ lib/Map/AbsConcept.js
