+++ lib/Models/AbsIttCatalogItem.js
@@ -16,7 +16,7 @@ var loadText = require('terriajs-cesium/Source/Core/loadText');
 var objectToQuery = require('terriajs-cesium/Source/Core/objectToQuery');
 var when = require('terriajs-cesium/Source/ThirdParty/when');
 
-var AbsConcept = require('./AbsConcept');
+var AbsConcept = require('../Map/AbsConcept');
 var arraysAreEqual = require('../Core/arraysAreEqual');
 var CsvCatalogItem = require('./CsvCatalogItem');
 var inherit = require('../Core/inherit');
